# docs/swagger.yaml
basePath: /
definitions:
  gin-scalable-api_pkg_response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_modules_branch.BranchListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_branch.BranchResponse'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  internal_modules_branch.BranchResponse:
    properties:
      code:
        type: string
      company_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      updated_at:
        type: string
    type: object
  internal_modules_branch.CreateBranchRequest:
    properties:
      code:
        maxLength: 20
        minLength: 2
        type: string
      company_id:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      parent_id:
        type: integer
    required:
    - code
    - company_id
    - name
    type: object
  internal_modules_branch.UpdateBranchRequest:
    properties:
      code:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Complete API for ERP with RBAC and API Documentation System. This API provides comprehensive endpoints for managing users, companies, branches, roles, modules, subscriptions, units, applications, and authentication with JWT-based security.

    ## Environments
    - **Development**: http://localhost:8081
    - **Staging**: https://staging-api.huminor.com
    - **Production**: https://api.huminor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Huminor Console API
  version: "1.0"
paths:
...
  /api/v1/branches:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua branch dengan filter opsional dan pagination.
        Mendukung nested hierarchy dengan query parameter nested=true
      parameters:
      - description: Limit jumlah data
        in: query
        name: limit
        type: integer
      - description: Offset data
        in: query
        name: offset
        type: integer
      - description: Search by name atau code
        in: query
        name: search
        type: string
      - description: Filter by company ID
        in: query
        name: company_id
        type: integer
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - description: Return nested hierarchy (true/false)
        in: query
        name: nested
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Branches berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_branch.BranchListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get all branches
      tags:
      - Branches
    post:
      consumes:
      - application/json
      description: Membuat branch baru dengan support untuk hierarchical structure
      parameters:
      - description: Branch data
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/internal_modules_branch.CreateBranchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Branch berhasil dibuat
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_branch.BranchResponse'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "409":
          description: Conflict - branch code sudah ada
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create new branch
      tags:
      - Branches
  /api/v1/branches/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus branch berdasarkan ID
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Branch berhasil dihapus
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid branch ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Branch tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Delete branch
      tags:
      - Branches
    get:
      consumes:
      - application/json
      description: Mendapatkan detail branch berdasarkan ID
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Branch berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_branch.BranchResponse'
              type: object
        "400":
          description: Bad request - Invalid branch ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Branch tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get branch by ID
      tags:
      - Branches
    put:
      consumes:
      - application/json
      description: Memperbarui informasi branch
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      - description: Branch data yang akan diupdate
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/internal_modules_branch.UpdateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Branch berhasil diupdate
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_branch.BranchResponse'
              type: object
        "400":
          description: Bad request - Invalid branch ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Branch tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update branch
      tags:
      - Branches
  /api/v1/branches/{id}/children:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar child branches dari branch tertentu
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Branch children berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_modules_branch.BranchResponse'
                  type: array
              type: object
        "400":
          description: Bad request - Invalid branch ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Branch tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get branch children
      tags:
      - Branches
  /api/v1/branches/{id}/hierarchy:
    get:
      consumes:
      - application/json
      description: Mendapatkan hierarchy branch berdasarkan ID. Mendukung nested format
        dengan query parameter nested=true
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      - description: Return nested hierarchy (true/false)
        in: query
        name: nested
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Branch hierarchy berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_branch.BranchResponse'
              type: object
        "400":
          description: Bad request - Invalid branch ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Branch tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get branch hierarchy
      tags:
      - Branches
...
  /api/v1/branches/company/{companyId}:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua branches dari company tertentu. Mendukung nested
        hierarchy dengan query parameter nested=true
      parameters:
      - description: Company ID
        in: path
        name: companyId
        required: true
        type: integer
      - description: Include hierarchy information (true/false)
        in: query
        name: include_hierarchy
        type: string
      - description: Return nested hierarchy (true/false)
        in: query
        name: nested
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Company branches berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_modules_branch.BranchResponse'
                  type: array
              type: object
        "400":
          description: Bad request - Invalid company ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Company tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get company branches
      tags:
      - Branches
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...